add autorizer method

+++ Control/CManageCompetition.php
@@ -1,16 +1,19 @@
 <?php
 
-require_once '../Foundation/FDbH.php';
-require_once '../Entity/EUser.php';
-require_once '../Entity/EEvent.php';
-
 class CManageCompetition
 {
+    private static function authorizer(ECompetition $competition):bool{
+
+        if(FSession::isLogged() && FDbH::loadEventByCompetition($competition)->getOrganizer()->getEmail()!=FSession::getUserLogged()->getEmail())throw new Exception("only the organizer can update competition");
+        return CManageUser::callLogin();
+    }
+
     public static function update(ECompetition $competition):void
     {
-        //VERIFICA LOGIN E TIPO UTENTE
         try{
-            if(!FDbH::updateOne($competition))throw new Exception("you can't update a competition that don't exist");
+            if(self::authorizer($competition)){
+                if(!FDbH::updateOne($competition))throw new Exception("you can't update a competition that don't exist");
+            }
         }
         catch (Exception $e){
             //RICHIAMA ERRORE
@@ -22,7 +25,9 @@ class CManageCompetition
         //VERIFICA LOGIN E TIPO UTENTE
         //if(filter_var($organizerEmail,FILTER_VALIDATE_EMAIL)!=false)
         try{
-            FDbH::store($competition);
+            if(self::authorizer($competition)){
+                FDbH::store($competition);
+            }
         }
         catch (Exception $e){
             //RICHIAMA ERRORE
@@ -31,7 +36,9 @@ class CManageCompetition
 
     public static function delete(ECompetition $competition){
         try{
-            FDbH::deleteOne($competition->getId(),ECompetition::class);
+            if(self::authorizer($competition)){
+                FDbH::deleteOne($competition->getId(),ECompetition::class);
+            }
         }
         catch(Exception $e){
                 //RICHIAMA ERRORE
@@ -49,8 +56,9 @@ class CManageCompetition
 
     public static function newPage(ECompetition $competition){
         try{
-            //VERIFICA LOGIN E TIPO UTENTE
-            //Richiama  VEvent::showNewPage($competition);
+            if(self::authorizer($competition)){
+                //Richiama  VEvent::showNewPage($competition);
+            }
         }
         catch(Exception $e){
             //RICHIAMA ERRORE
